==> STEP01_SPE.ipynb <==
# SMILES Pair Encoding

In this Jupyter notebook, the vocabulary is generated by SMILES Pair Encoding.
With my computing resources, this step takes about 1.5 hours.

<br>

### File Requirements

The following files are required in the `data/STEP01` folder:
1. [ChEMBL_SMILES_2kk.csv](https://www.ebi.ac.uk/chembl/explore/compounds/)

If you download the ChEMBL SMILES strings (approx. 140MB) again, it is possible that the SMILES Pair Encoder will throw an error message, as all SMILES strings must be kekulizable.
Filtering the SMILES strings with the “Chem” module from rdkit can help here.

<br>

### Output
The trained vocabulary “SPE_ChEMBL_1500freq.txt” needed in "STEP01_SMILES_ENCODER.ipynb".

import codecs
from SmilesPE.learner import *
import pandas as pd 

df = pd.read_csv("data/STEP01/ChEMBL_SMILES_2kk.csv")
smiles_list = df.values.tolist()

output = codecs.open("data/STEP01/SPE_ChEMBL_1500freq.txt", "w")
learn_SPE(smiles_list, output, 30000, min_frequency=1500, augmentation=1, verbose=True, total_symbols=True)
